+++ pytsite/cache/_api.py
@@ -2,7 +2,7 @@
 """
 from typing import Dict as _Dict
 from pytsite import logger as _logger, reg as _reg
-from . import _driver
+from . import _driver, _error
 
 __author__ = 'Alexander Shepetko'
 __email__ = 'a@shepetko.com'
@@ -12,72 +12,72 @@ _pools = {}  # type: _Dict[str, _driver.Abstract]
 _dbg = _reg.get('cache.debug')
 
 
-def has_pool(name: str) -> bool:
+def has_pool(uid: str) -> bool:
     """Check whether a pool exists.
     """
-    return name in _pools
+    return uid in _pools
 
 
-def create_pool(name: str, driver: str = None) -> _driver.Abstract:
+def create_pool(uid: str, driver: str = None) -> _driver.Abstract:
     """Create new pool.
     """
-    if name in _pools:
-        raise KeyError("Cache pool '{}' already exists.".format(name))
+    if uid in _pools:
+        raise _error.PoolExists("Cache pool '{}' already exists.".format(uid))
 
     if driver is None:
         driver = _reg.get('cache.driver', 'redis')
 
     if driver == 'redis':
-        drv = _driver.Redis(name)
+        drv = _driver.Redis(uid)
     else:
-        raise ValueError("Cache driver '{}' is not supported.".format(driver))
+        raise _error.DriverNotFound("Cache driver '{}' is not supported.".format(driver))
 
-    _pools[name] = drv
+    _pools[uid] = drv
 
     if _dbg:
-        _logger.debug("POOL CREATED: '{}', driver: '{}'.".format(name, driver), __name__)
+        _logger.debug("POOL CREATED: '{}', driver: '{}'.".format(uid, driver), __name__)
 
     return drv
 
 
-def get_pool(name: str) -> _driver.Abstract:
+def get_pool(uid: str) -> _driver.Abstract:
     """Get a pool.
     """
-    if name not in _pools:
-        raise KeyError("Pool '{}' is not defined.".format(name))
+    if uid not in _pools:
+        raise _error.PoolNotExist("Pool '{}' does not exist.".format(uid))
 
     if _dbg:
-        _logger.debug("POOL GET: '{}'.".format(name), __name__)
+        _logger.debug("POOL GET: '{}'.".format(uid), __name__)
 
-    return _pools[name]
+    return _pools[uid]
 
 
-def clear_pool(name: str):
+def clear_pool(uid: str):
     """Clear a pool.
     """
-    if name not in _pools:
-        raise KeyError("Pool '{}' is not defined.".format(name))
+    if uid not in _pools:
+        raise KeyError("Pool '{}' is not defined.".format(uid))
 
-    _pools[name].clear()
+    _pools[uid].clear()
 
 
-def delete_pool(name: str):
+def delete_pool(uid: str):
     """Delete a pool.
     """
-    clear_pool(name)
-    del _pools[name]
+    clear_pool(uid)
+    del _pools[uid]
 
     if _dbg:
-        _logger.debug("POOL DELETE: '{}'.".format(name), __name__)
+        _logger.debug("POOL DELETE: '{}'.".format(uid), __name__)
 
 
-def cleanup_pool(name: str):
+def cleanup_pool(uid: str):
     """Clear expired items in a pool.
     """
     if _dbg:
-        _logger.debug("POOL CLEANUP: {}.".format(name), __name__)
+        _logger.debug("POOL CLEANUP: {}.".format(uid), __name__)
 
-    get_pool(name).cleanup()
+    get_pool(uid).cleanup()
 
 
 def cleanup_pools():