+++ pytsite/cache/_error.py
@@ -6,5 +6,17 @@ __email__ = 'a@shepetko.com'
 __license__ = 'MIT'
 
 
+class DriverNotFound(Exception):
+    pass
+
+
+class PoolNotExist(Exception):
+    pass
+
+
+class PoolExists(Exception):
+    pass
+
+
 class KeyNotExist(Exception):
     pass