+++ pytsite/odm/__init__.py
@@ -24,5 +24,6 @@ def __init():
 
     # Reindex collections on every update
     events.listen('pytsite.update.after', _eh.pytsite_update_after)
+    events.listen('pytsite.db.restore', _eh.pytsite_db_restore)
 
 __init()