+++ pytsite/odm/_eh.py
@@ -1,5 +1,15 @@
-from pytsite import console as _console
+from pytsite import console as _console, lang as _lang
+from . import _api
 
 
 def pytsite_update_after():
     _console.run_command('odm', reindex=True)
+
+
+def pytsite_db_restore():
+    _api.get_entities_cache().clear()
+    _console.print_info(_lang.t('pytsite.odm@entities_cache_cleared'))
+
+    for model in _api.get_registered_models():
+        _api.get_finder_cache(model).clear()
+        _console.print_info(_lang.t('pytsite.odm@finder_cache_cleared', {'model': model}))